+++ src/coffeemachineapp2/Grinder.java
@@ -11,15 +11,15 @@ import coffeemachineapp2.Exceptions.GrinderLevelException;
  * @author abdsh
  */
 public class Grinder {
-    private int grindLevel;
-    private int beansNedded;
+    private double grindLevel;
+    private double beansNedded;
 private FileLogger logger;
 
     public Grinder(FileLogger logger) {
         this.logger = logger;
     }
     
-    public void grind(int grindLevel){
+    public void grind(double grindLevel){
         if(grindLevel==1)
             logger.log("Grind level set to  1 (Coarse).");
         else if(grindLevel==2)
@@ -32,22 +32,22 @@ private FileLogger logger;
                            logger.log("Grind level set to 5 (Very fine)");
     }
 
-    public int getBeansNedded() {
+    public double getBeansNedded() {
         return beansNedded;
     }
 
-    public void setBeansNedded(int beansNedded) {
+    public void setBeansNedded(double beansNedded) {
         this.beansNedded = beansNedded;
     }
 
-    public void setGrindSize(int grindSize) {
+    public void setGrindSize(double grindSize) {
         this.grindLevel = grindSize;
     }
 
     
     
 
-    public int getGrindSize() {
+    public double getGrindSize() {
         return grindLevel;
     }
 }
